requirements.txt
Selenium==4.33.0
pytest==8.4.0
webdriver-manager==4.0.2
pytest-order==1.3.0
python-dotenv==1.1.0